Converted most of LoginServer to ReadPacket and std::string
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@503 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/Common/ReadPacket.cpp
@@ -23,7 +23,7 @@ unsigned char ReadPacket::getByte() {
 	return buffer[pos++];
 }
 
-void ReadPacket::skipBytes(size_t len) {
+void ReadPacket::skipBytes(int len) {
 	pos += len;
 }
